Singleton objects were not singletons when made immutable. We need to provide a custom Meta::Method::Constructor so the generated immtable new() method respects singleton-ness.
This test implements this, along with tests.

Finally, it adds no Moose at the end of all the classes to prevent any
conflicts between Moose's exports and the class's methods. This was a
problem in Meta::Class, where the exported make_immutable was
overriding the one provided by Moose::Meta::Class.